Refactor: relocate description-parsing logic
